Move network status to status bar
