+++ src/plugins/console/public/application/containers/editor/monaco/monaco_editor_output.tsx
@@ -92,6 +92,7 @@ export const MonacoEditorOutput: FunctionComponent = () => {
     <div
       css={css`
         width: 100%;
+        height: 100%;
       `}
       ref={divRef}
     >
