+++ src/plugins/console/public/styles/_app.scss
@@ -123,10 +123,6 @@
   z-index: $euiZLevel2;
 }
 
-.conApp__tabsExtension {
-  border-bottom: $euiBorderThin;
-}
-
 /*
  * The highlighting for the selected requests in the monaco editor
  */
